Pass tempfile param when building archive files                         -
Because the archive file object has a lazy tempdir attribute everytime
when an archive file is built and a temdir is invoked a new temporary
directory is created. This causes problems for books with many pages.
Image conversion operations are failing in an unpredictable manner.

Because the archive file object has a lazy tempdir attribute everytime
when an archive file is built and a temdir is invoked a new temporary
directory is created. This causes problems for books with many pages.
Image conversion operations are failing in an unpredictable manner.